Allow chargers to provide vehicle SoC via ISO15118
Some chargers (e.g. Porsche/Audi Mobile Charger Connect) can provide the EVs current SoC via the ISO15118 connection to the car.

This change allows chargers to implement the SoC method and return the SoC or api.ErrNotAvailable if the currently connected car does not provide the data.

As the SoC via this connection is always up to date, the estimator does not need to run any more